WIP: Continue experiments ..

+++ p2panda-rs/src/operation/operation.rs
@@ -76,21 +76,20 @@ impl Copy for OperationAction {}
 #[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
 #[serde(tag = "type", content = "value")]
 pub enum OperationValue {
-    /// Basic `boolean` value.
+    /// Boolean value.
     #[serde(rename = "bool")]
     Boolean(bool),
 
-    /// Basic signed `integer` value.
+    /// Signed integer value.
     #[serde(rename = "int")]
     #[serde(with = "serde_with::rust::display_fromstr")]
-    Integer(u64),
+    Integer(i64),
 
-    /// Basic signed `float` value.
+    /// Floating point value.
     #[serde(rename = "float")]
-    #[serde(with = "serde_with::rust::display_fromstr")]
     Float(f64),
 
-    /// Basic `string` value.
+    /// String value.
     #[serde(rename = "str")]
     Text(String),
